Move accessibility test helper to test helpers file
