+++ src/Hedwig/ClientApp/src/containers/Reports/ReportDetail/UtilizationTable.test.tsx
@@ -4,7 +4,7 @@ import { Age, FundingTime, CdcReport, FundingSource, Enrollment, Region } from '
 import UtilizationTable, { calculateRate } from './UtilizationTable';
 import emptyGuid from '../../../utils/emptyGuid';
 import cartesianProduct from '../../../utils/cartesianProduct';
-import { accessibilityTestHelper } from '../../accessibilityTestHelper';
+import { accessibilityTestHelper } from '../../../tests/helpers';
 
 describe('calculateRate', () => {
 	it('includes all possible rates', () => {
