+++ src/Hedwig/ClientApp/src/containers/Reports/ReportsSummary/ReportsSummary.test.tsx
@@ -23,7 +23,7 @@ import React from 'react';
 import { render } from '@testing-library/react';
 import ReportsSummary from './ReportsSummary';
 import CommonContextProviderMock from '../../../contexts/__mocks__/TestProvider';
-import { accessibilityTestHelper } from '../../accessibilityTestHelper';
+import { accessibilityTestHelper } from '../../../tests/helpers';
 
 afterAll(() => {
 	jest.resetModules();
